Удаление попапа при открытии другого DOM
